Added enemy aminations for idle, walking and attacking
notes: 8 questions. worked surprisingly well - likely due to lack of animations, not as complex as player ones. Only negative was buttons not where it thought, outdated info. Weird thing: tried to give me images from "imgur.com" this time for the animator controller but the images didnt work and say "the image you are requesting does not exist or is no longer available."