enemy health bar working as intended
Notes: To get base functionality 3 questions. To make it be above the enemies head and move with it, 19 and it had failed with 1 reset at 14 and then it told me to do the same thing again but made a small separate health bar script to get base functionality in with only 1 question. It kept trying to instantiate the health bar in and used WorldToScreenPoint for the location. After strong prompt of "i think instead of spawning in the health bar we should just attach it to the enemy prefab" it completed in 2, second question was to remove "WorldToScreenPoint" as i told it that if i include that line it goes miles away but if i dont then it works perfectly.